Adds permalink test (currently failing)
