+++ test/stubs/class-data-permalink.11ty.js
@@ -1,0 +1,14 @@
+class Test {
+  get data() {
+    return {
+      permalink: "/my-permalink/"
+    };
+  }
+  // TODO permalink: function() {}
+
+  render({ name }) {
+    return `<p>${name}</p>`;
+  }
+}
+
+module.exports = Test;
